Remove redundant code in SampleDataUtil
