+++ src/test/java/seedu/address/testutil/PatientBuilder.java
@@ -14,7 +14,7 @@ import seedu.address.model.person.patient.Remark;
 import seedu.address.model.person.patient.Status;
 import seedu.address.model.person.patient.Weight;
 import seedu.address.model.tag.Tag;
-import seedu.address.model.util.SampleDataUtil;
+import seedu.address.model.util.LoaderUtil;
 
 /**
  * A utility class to help with building {@code Patient} objects.
@@ -86,7 +86,7 @@ public class PatientBuilder {
      * Parses the {@code tags} into a {@code Set<Tag>} and set it to the {@code Patient} that we are building.
      */
     public PatientBuilder withTags(String ... tags) {
-        this.tags = SampleDataUtil.getTagSet(tags);
+        this.tags = LoaderUtil.getTagSet(tags);
         return this;
     }
