+++ src/test/java/seedu/address/testutil/PersonBuilder.java
@@ -8,7 +8,7 @@ import seedu.address.model.person.Name;
 import seedu.address.model.person.Person;
 import seedu.address.model.person.Phone;
 import seedu.address.model.tag.Tag;
-import seedu.address.model.util.SampleDataUtil;
+import seedu.address.model.util.LoaderUtil;
 
 /**
  * A utility class to help with building Person objects.
@@ -56,7 +56,7 @@ public class PersonBuilder {
      * Parses the {@code tags} into a {@code Set<Tag>} and set it to the {@code Person} that we are building.
      */
     public PersonBuilder withTags(String ... tags) {
-        this.tags = SampleDataUtil.getTagSet(tags);
+        this.tags = LoaderUtil.getTagSet(tags);
         return this;
     }
